buffer: globally refactor buffer creation
