+++ src/core/abr/representation_chooser.ts
@@ -49,7 +49,7 @@ const {
 interface IRepresentationChooserClockTick {
   bitrate : number|undefined; // currently set bitrate, in bit per seconds
   bufferGap : number; // time to the end of the buffer, in seconds
-  position : number; // current position, in seconds
+  currentTime : number; // current position, in seconds
   speed : number; // current playback rate
 }
 
@@ -238,7 +238,7 @@ function estimateStarvationModeBitrate(
   clock : IRepresentationChooserClockTick,
   lastEstimatedBitrate : number|undefined
 ) : number|undefined {
-  const nextNeededPosition = clock.position + clock.bufferGap;
+  const nextNeededPosition = clock.currentTime + clock.bufferGap;
   const concernedRequest = getConcernedRequest(requests, nextNeededPosition);
   if (!concernedRequest) {
     return undefined;
