+++ src/core/buffer/active_period_emitter.ts
