+++ src/core/buffer/segment_bookkeeper.ts
