+++ src/core/buffer/segment_filter.ts
@@ -22,7 +22,7 @@ import {
   Representation,
 } from "../../manifest";
 import SimpleSet from "../../utils/simple_set";
-import { SegmentBookkeeper } from "../stream";
+import SegmentBookkeeper from "./segment_bookkeeper";
 
 const {
   BITRATE_REBUFFERING_RATIO,
