+++ src/core/stream/create_eme_manager.ts
@@ -51,3 +51,5 @@ export default function createEMEManager(
 
   return features.emeManager(mediaElement, keySystems);
 }
+
+export { IEMEManagerEvent };
