+++ src/core/stream/on_live_buffer_event.ts
@@ -26,11 +26,10 @@ import {
 } from "rxjs/operators";
 import log from "../../log";
 import Manifest from "../../manifest";
+import { IPeriodBufferManagerEvent } from "../buffer";
 import SourceBufferManager from "../source_buffers";
-import { IBufferHandlerEvent } from "./buffers_handler";
-import EVENTS, {
-  IManifestUpdateEvent,
-} from "./stream_events";
+import EVENTS from "./events_generators";
+import { IManifestUpdateEvent } from "./types";
 
 /**
  * Re-fetch the manifest and merge it with the previous version.
@@ -60,7 +59,7 @@ function refreshManifest(
 
 export type ILiveEventsHandlerEvent =
   IManifestUpdateEvent |
-  IBufferHandlerEvent;
+  IPeriodBufferManagerEvent;
 
 /**
  * Create handler for Buffer events happening only in live contexts.
@@ -73,7 +72,7 @@ export default function liveEventsHandler(
   videoElement : HTMLMediaElement,
   manifest : Manifest,
   fetchManifest : (url : string) => Observable<Manifest>
-) : (message : IBufferHandlerEvent) => Observable<ILiveEventsHandlerEvent> {
+) : (message : IPeriodBufferManagerEvent) => Observable<ILiveEventsHandlerEvent> {
   return function handleLiveEvents(message) {
     switch (message.type) {
       case "discontinuity-encountered":
